Refactor to use etcd client library
