+++ governor.py
@@ -3,7 +3,7 @@
 import sys, os, yaml, time, urllib2, atexit
 import logging
 
-from helpers.etcd import Etcd
+from helpers.keystore import Etcd
 from helpers.postgresql import Postgresql
 from helpers.ha import Ha
 
@@ -62,8 +62,8 @@ while True:
 
     # create replication slots
     if postgresql.is_leader():
-        for node in etcd.get_client_path("/members?recursive=true")["node"]["nodes"]:
-            member = node["key"].split('/')[-1]
+        for member in etcd.members():
+            member =  member['hostname']
             if member != postgresql.name:
                 postgresql.query("DO LANGUAGE plpgsql $$DECLARE somevar VARCHAR; BEGIN SELECT slot_name INTO somevar FROM pg_replication_slots WHERE slot_name = '%(slot)s' LIMIT 1; IF NOT FOUND THEN PERFORM pg_create_physical_replication_slot('%(slot)s'); END IF; END$$;" % {"slot": member})
